+++ examples/simpledb.lua
@@ -22,5 +22,6 @@ skynet.start(function()
 			error(string.format("Unknown command %s", tostring(cmd)))
 		end
 	end)
+	skynet.register ".simpledb"
 	skynet.register "SIMPLEDB"
 end)