+++ lualib/cluster.lua
@@ -4,7 +4,7 @@ local clusterd
 local cluster = {}
 
 function cluster.call(node, address, ...)
-	-- skynet.pack(...) will free by cluster.c.packrequest
+	-- skynet.pack(...) will free by cluster.core.packrequest
 	return skynet.call(clusterd, "lua", "req", node, address, skynet.pack(...))
 end
 
@@ -18,6 +18,23 @@ end
 
 function cluster.reload()
 	skynet.call(clusterd, "lua", "reload")
+end
+
+function cluster.proxy(node, name)
+	return skynet.call(clusterd, "lua", "proxy", node, name)
+end
+
+local namecache = {}
+
+function cluster.ncall(name, ...)
+	local s = namecache[name]
+	if not s then
+		local node , lname = name:match "(.-)(%..+)"
+		assert(node and lname)
+		s = cluster.proxy(node, lname)
+		namecache[name] = assert(s)
+	end
+	return skynet.call(s, "lua", ...)
 end
 
 skynet.init(function()